packages/ml-core/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.0.0
pytest>=8.0.0 